Animations Working in game
BERN SHOW-TAN-ES

+++ Bern-Work-Sprint-1/Trinity Solution/source/Animation/CFrame.h
@@ -12,7 +12,7 @@ class CFrame
 private:
 	RECT m_DrawRect, m_CollisionRect;
 	POINT m_AnchorPoint;
-	float m_fDuration;
+	double m_fDuration;
 	string m_szEvent;
 	
 public:
@@ -20,13 +20,13 @@ public:
 	RECT GetDrawRect() const { return m_DrawRect; }
 	RECT GetCollisionRect() const { return m_CollisionRect; }
 	POINT GetAnchorPoint() const { return m_AnchorPoint; }
-	float GetDuration() const { return m_fDuration; }
+	double GetDuration() const { return m_fDuration; }
 	string GetEvent() const { return m_szEvent; }
 	//	Mutators
 	void SetDrawRect(RECT val) { m_DrawRect = val; }
 	void SetCollisionRect(RECT val) { m_CollisionRect = val; }
 	void SetAnchorPoint(POINT val) { m_AnchorPoint = val; }
-	void SetDuration(float val) { m_fDuration = val; }
+	void SetDuration(double val) { m_fDuration = val; }
 	void SetEvent(string val) { m_szEvent = val; }
 };
 #endif
